# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WorkForcePro API",
    "version": "1.0.0",
    "description": "RESTful API for managing employees, trainings, certifications, and departments in a workforce management system."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/employees": {
      "get": {
        "tags": ["Employees"],
        "summary": "List all employees",
        "operationId": "getEmployees",
        "responses": {
          "200": {
            "description": "List of employees with department and supervisor info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Employee" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Employees"],
        "summary": "Create a new employee",
        "operationId": "createEmployee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEmployeeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid supervisor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/employees/supervisors": {
      "get": {
        "tags": ["Employees"],
        "summary": "List all supervisors",
        "operationId": "getSupervisors",
        "responses": {
          "200": {
            "description": "List of employees with supervisor privileges",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Supervisor" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/employees/{id}": {
      "get": {
        "tags": ["Employees"],
        "summary": "Get a single employee",
        "operationId": "getEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "200": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "tags": ["Employees"],
        "summary": "Update an employee",
        "operationId": "updateEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateEmployeeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Employee" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Employees"],
        "summary": "Delete an employee",
        "operationId": "deleteEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "204": { "description": "Employee deleted, no content" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/employees/{id}/trainings": {
      "get": {
        "tags": ["Employees", "Trainings"],
        "summary": "Get trainings for an employee",
        "operationId": "getEmployeeTrainings",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "200": {
            "description": "Employee training assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EmployeeTraining" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/employees/{id}/certifications": {
      "get": {
        "tags": ["Employees", "Certifications"],
        "summary": "Get certifications for an employee",
        "operationId": "getEmployeeCertifications",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "200": {
            "description": "Employee certifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/trainings/employee/{employee_id}": {
      "get": {
        "tags": ["Trainings"],
        "summary": "Get trainings for a specific employee",
        "operationId": "getTrainingsByEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeNumericId" }
        ],
        "responses": {
          "200": {
            "description": "Employee training assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EmployeeTraining" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/certifications/employee/{employee_id}": {
      "get": {
        "tags": ["Certifications"],
        "summary": "Get certifications for a specific employee",
        "operationId": "getCertificationsByEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeNumericId" }
        ],
        "responses": {
          "200": {
            "description": "Employee certifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Certification" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "EmployeeId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
      "EmployeeNumericId": {
        "name": "employee_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
...
    },
    "schemas": {
...
      "Supervisor": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "employee_id": { "type": "string", "example": "EMP001" },
          "name": { "type": "string", "example": "Jane Smith" },
          "email": { "type": "string", "example": "jane.smith@company.com" }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "employee_id": { "type": "string", "example": "EMP001" },
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "example": "john.doe@company.com" },
          "department_id": { "type": "integer", "nullable": true, "example": 2 },
          "department_name": { "type": "string", "nullable": true, "example": "Engineering" },
          "supervisor_id": { "type": "integer", "nullable": true, "example": 3 },
          "supervisor_name": { "type": "string", "nullable": true, "example": "Jane Smith" },
          "is_supervisor": { "type": "boolean", "example": false },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateEmployeeRequest": {
        "type": "object",
        "required": ["employee_id", "name", "email"],
        "properties": {
          "employee_id": { "type": "string", "example": "EMP010" },
          "name": { "type": "string", "example": "Alice Johnson" },
          "email": { "type": "string", "format": "email", "example": "alice.johnson@company.com" },
          "department_id": { "type": "integer", "nullable": true, "example": 1 },
          "supervisor_id": { "type": "integer", "nullable": true, "example": 3 },
          "is_supervisor": { "type": "boolean", "example": false }
        }
      },
      "UpdateEmployeeRequest": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": { "type": "string", "example": "Alice Johnson-Smith" },
          "email": { "type": "string", "format": "email", "example": "alice.johnson@company.com" },
          "department_id": { "type": "integer", "nullable": true, "example": 2 },
          "supervisor_id": { "type": "integer", "nullable": true, "example": 3 },
          "is_supervisor": { "type": "boolean", "example": true }
        }
      },
...
      "EmployeeTraining": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "employee_id": { "type": "integer", "example": 2 },
          "training_id": { "type": "integer", "example": 1 },
          "training_name": { "type": "string", "example": "Workplace Safety" },
          "due_date": { "type": "string", "format": "date", "nullable": true, "example": "2026-07-25" },
          "status": {
            "type": "string",
            "enum": ["assigned", "completed", "approved", "rejected"],
            "example": "assigned"
          },
          "completion_date": { "type": "string", "format": "date", "nullable": true, "example": "2026-07-20" },
          "notes": { "type": "string", "nullable": true, "example": "Completed all modules" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Certification": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "employee_id": { "type": "integer", "example": 2 },
          "employee_name": { "type": "string", "example": "John Doe" },
          "certification_name": { "type": "string", "example": "AWS Solutions Architect" },
          "expiry_date": { "type": "string", "format": "date", "example": "2026-12-31" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Employee not found" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}